feat: display search input box in h5 env
在h5环境显示搜索输入框

Log:
